add mapping features for four classes and changed long balance/cost values to integer

+++ src/main/java/com/codethecode/courseregistersystem/dto/CourseDto.java
@@ -9,5 +9,5 @@ public class CourseDto {
     private String day;
     private String teacherId;
     private String studentId;
-    private Long cost;
+    private Integer cost;
 }
